Fix build failures in Aquarium ANGLE backend
This patch fixes the build failures in Aquarium ANGLE backend by
updating ANGLE to the latest revision which has the required fix.

Note that we still need to specify "angle_enable_vulkan = false"
because there are still conflicts between SPIRV-Tools in Dawn and
ANGLE. We will try to fix this issue in the future.